#45: Work on tests

+++ gslab_scons/tests/test_release_function.py
@@ -5,7 +5,9 @@ import re
 import mock
 import shutil
 import requests
+import copy
 # Import module containing gslab_scons testing side effects
+import _test_helpers as helpers
 import _side_effects as fx
 
 # Ensure that Python can find and load the GSLab libraries
@@ -15,9 +17,8 @@ sys.path.append('../..')
 import gslab_scons
 import gslab_scons._release_tools as tools
 from gslab_scons._exception_classes import ReleaseError
-from gslab_make.tests import nostderrout
 
-
+#== Preliminaries ===================================================
 # Create a patch for use in testing release()
 path = 'gslab_scons._release_tools'
 patch_collection = \
@@ -34,6 +35,19 @@ patch_collection = \
                 f \
               ))))))))))
 
+# Standard release() arguments
+standard_args = {'vers':              'test_version',
+                 'DriveReleaseFiles': ['paper.pdf', 'plot.pdf'],
+                 'local_release':     'root/release/repo-test_branch',
+                 'org':               'org',
+                 'repo':              'repo',
+                 'target_commitish':  'test_branch',
+                 'zip_release':       True}
+
+# This is mocked-up output from requests.session.get(path) based on 
+# members of standard_args
+org_repo = (standard_args['org'], standard_args['repo'])
+#====================================================================
 
 class TestReleaseFunction(unittest.TestCase):
 
@@ -55,8 +69,11 @@ class TestReleaseFunction(unittest.TestCase):
         mock_getpass.return_value = 'test_token'
         mock_session.return_value = mock.MagicMock(post = mock.MagicMock(),
                                                    get  = mock.MagicMock())
-        mock_session.return_value.get.return_value.content = \
-            self.return_mock_release_data()
+
+        args = copy.deepcopy(standard_args)
+        release_data = helpers.mock_release_data(args)
+        mock_session.return_value.get.return_value.content = release_data
+
         # When org/repo refers to a nonexistent GitHub repository, 
         # make the release request fail. 
         mock_session.return_value.post.side_effect = fx.post_side_effect
@@ -64,15 +81,25 @@ class TestReleaseFunction(unittest.TestCase):
         # Give mock_upload a side effect preserving release assets listing
         mock_upload.side_effect = fx.upload_asset_side_effect
 
-        DriveReleaseFiles = ['paper.pdf', 'plot.pdf']
-        local = 'Google Drive/release/local_release/'
-        tools.release('test_version', 
-                      DriveReleaseFiles = DriveReleaseFiles, 
-                      local_release     = local, 
-                      org               = 'org', 
-                      repo              = 'repo', 
-                      target_commitish  = 'test_branch', 
-                      zip_release       = True)
+        args['zip_release'] = True
+
+        self.check_release(args, mock_session, mock_upload, mock_copy,
+                           mock_make_archive, mock_move)
+
+        # Test release() when DriveReleaseFiles is a string
+        args['DriveReleaseFiles'] = 'plot.pdf'
+        self.check_release(args, mock_session, mock_upload, mock_copy,
+                           mock_make_archive, mock_move)
+
+
+    def check_release(self, args, mock_session, mock_upload, mock_copy, 
+                      mock_make_archive, mock_move):
+        
+        for mock_object in [mock_session, mock_upload, mock_copy,
+                           mock_make_archive, mock_move]:
+            mock_object.reset_mock()
+
+        tools.release(**args)
 
         mock_session.return_value.post.assert_called_once()
         post_args = mock_session.return_value.post.call_args
@@ -86,9 +113,10 @@ class TestReleaseFunction(unittest.TestCase):
         # Check that the correct data was passed to session.post()
         data = post_args[1]['data']
         self.assertTrue(re.search('"body": ""', data))
-        self.assertTrue(re.search('"name": "test_version"', data))
-        self.assertTrue(re.search('"target_commitish": "test_branch"', data))
-        self.assertTrue(re.search('"tag_name": "test_version"', data))
+        self.assertTrue(re.search('"name": "%s"' % args['vers'], data))
+        self.assertTrue(re.search('"target_commitish": "%s"' % \
+                                  args['target_commitish'], data))
+        self.assertTrue(re.search('"tag_name": "%s"' % args['vers'], data))
         self.assertTrue(re.search('"prerelease": false', data))
         self.assertTrue(re.search('"draft": false', data))
 
@@ -98,47 +126,37 @@ class TestReleaseFunction(unittest.TestCase):
         # of session.get()
         self.assertEqual(mock_upload.call_args[1]['release_id'], 'test_ID')
 
-        # Check that release() zipped the files released to Google Drive correctly...
-        mock_copy.assert_any_call('paper.pdf', 'release_content/paper.pdf')
-        mock_copy.assert_any_call('plot.pdf',  'release_content/plot.pdf')
-        mock_make_archive.assert_called_with('release_content', 'zip', 
-                                             'release_content')
+        # Check that release() prepared files for release correctly,
+        if isinstance(args['DriveReleaseFiles'], list):
+            for filename in args['DriveReleaseFiles']:
+                mock_copy.assert_any_call(filename, 
+                                          'release_content/%s' % filename)
+        else:
+            filename = args['DriveReleaseFiles']
+            mock_copy.assert_any_call(filename, 
+                                      'release_content/%s' % filename)
+        # ...that it zipped them if zip_release == True, ...
+        if args['zip_release']:
+            mock_make_archive.assert_called_with('release_content', 'zip', 
+                                                 'release_content')
+        else:
+            mock_make_archive.assert_not_called()
 
-        # ...and moved them to the local_release Google Drive directory
+        # ...and that it moved them to the local_release directory
         mock_move.assert_called_with('release_content.zip', 
-                                     'Google Drive/release/local_release/release.zip')
+                                     '%s/release.zip' % args['local_release'])
         # Check that the assets listed in the file whose path is
         # passed to upload_asset() are those specified by release()'s 
         # DriveReleaseFiles argument.
-        with open('assets_listing.txt', 'rU') as assets:
-            lines = assets.readlines()
-        self.assertIn(lines[1].strip(), DriveReleaseFiles)
-        self.assertIn(lines[2].strip(), DriveReleaseFiles)
+        if args['DriveReleaseFiles']:
+            with open('assets_listing.txt', 'rU') as assets:
+                lines = assets.readlines()
+    
+            for line in lines:
+                if not line == lines[0]:
+                    self.assertIn(line.strip(), args['DriveReleaseFiles'])
 
-        os.remove('assets_listing.txt')
-
-    @staticmethod
-    def return_mock_release_data():
-        '''This is mocked-up output from session.get(path)'''
-        return  ','.join([
-                '[{"url":'
-                    '"https://api.github.com/'
-                    'repos/org/repo/releases/test_ID"',
-                 '"assets_url":'
-                    'https://api.github.com/'
-                    'repos/org/repo/releases/test_ID/assets"',
-                 '"upload_url":'
-                    '"https://uploads.github.com/'
-                    'repos/org/repo/releases/test_ID/assets{?name,label}"',
-                 '"html_url":'
-                    '"https://github.com/'
-                    'org/repo/releases/tag/test_version"',
-                 '"id":test_ID',
-                 '"tag_name":"test_version"',
-                 '"target_commitish":"test_branch"',
-                 '"name":"test_version"',
-                 '"draft":false',
-                 '"prerelease":false}]'])
+            os.remove('assets_listing.txt')
 
     @patch_collection
     def test_release_nozip(self,
@@ -156,24 +174,14 @@ class TestReleaseFunction(unittest.TestCase):
         '''
         # Mock functions called by release() to simulate an actual call
         mock_getpass.return_value = 'test_token'
-        mock_session.return_value.get.return_value.content = self.return_mock_release_data()
+        args = copy.deepcopy(standard_args)
 
-        DriveReleaseFiles = ['paper.pdf', 'plot.pdf']
-        local = 'Google Drive/release/local_release'
-        tools.release('test_version', 
-                      DriveReleaseFiles = DriveReleaseFiles, 
-                      local_release     = local, 
-                      org               = 'org', 
-                      repo              = 'repo', 
-                      target_commitish  = 'test_branch', 
-                      zip_release       = False)
-        # The release() call should not have created a zip archive...
-        mock_make_archive.assert_not_called()
-        # ...but it should have moved the release files into the 
-        # local_release directory.
-        local = 'Google Drive/release/local_release'
-        mock_copy.assert_any_call('paper.pdf', '%s/paper.pdf' % local)
-        mock_copy.assert_any_call('plot.pdf',  '%s/plot.pdf'  % local)
+        release_data = helpers.mock_release_data(args)
+        mock_session.return_value.get.return_value.content = release_data
+
+        args['zip_release'] = False
+        self.check_release(args, mock_session, mock_upload, mock_copy,
+                           mock_make_archive, mock_move)
 
     @patch_collection
     def test_release_no_drive(self,
@@ -190,16 +198,14 @@ class TestReleaseFunction(unittest.TestCase):
         that does not upload files to Google Drive.
         '''
         # Mock functions called by release() to simulate an actual call
-        mock_session.return_value.get.return_value.content = \
-            self.return_mock_release_data()
+        args = copy.deepcopy(standard_args)
+        args['DriveReleaseFiles'] = []
+
+        release_data = helpers.mock_release_data(args)
+        mock_session.return_value.get.return_value.content = release_data
 
         # Test without DriveReleaseFiles
-        tools.release('test_version', 
-                      DriveReleaseFiles = [], 
-                      local_release     = 'Google Drive/release/local_release/', 
-                      org               = 'org', 
-                      repo              = 'repo', 
-                      target_commitish  = 'test_branch')
+        tools.release(**args)
         # Check that no file operations occur when no files are specified for release
         # to Google Drive.
         mock_copy.assert_not_called()      
@@ -225,80 +231,60 @@ class TestReleaseFunction(unittest.TestCase):
         '''
         # Mock functions called by release() to simulate an actual call
         mock_getpass.return_value = 'test_token'
-        mock_session.return_value.get.return_value.content = \
-            self.return_mock_release_data()
+        args = copy.deepcopy(standard_args)
+
+        release_data = helpers.mock_release_data(args)
+        mock_session.return_value.get.return_value.content = release_data
         mock_session.return_value.post.side_effect = fx.post_side_effect
 
-        # Inappropriate local_release argument should not if the DriveReleaseFiles
-        # argument is empty.
+        args = copy.deepcopy(standard_args)
+
+        # Inappropriate local_release argument and DriveReleaseFiles is empty.
+        test_args = copy.deepcopy(args)
+        test_args['DriveReleaseFiles'] = []
+        test_args['local_release']     = 1
         try:
-            tools.release('test_version', 
-                          DriveReleaseFiles = [], 
-                          local_release     = 1, 
-                          org               = 'org', 
-                          repo              = 'repo', 
-                          target_commitish  = 'test_branch')
+            tools.release(**test_args)
         except:
-            self.fail("Running release() with an invalid local_release argument and without"
-                      "files to be released to Google Drive shouldn't raise an error!")
+            self.fail("Running release() with an invalid local_release "
+                      "argument and without files to be released to Google "
+                      "Drive shouldn't raise an error!")
 
-        # Providing a local_release argument of an inappropriate type should
-        # raise an error when we instruct release() to upload files to Google Drive
-        with self.assertRaises(AttributeError), nostderrout():
-            tools.release('test_version', 
-                          DriveReleaseFiles = ['paper.pdf', 'plot.pdf'], 
-                          local_release     = 1, 
-                          org               = 'org', 
-                          repo              = 'repo', 
-                          target_commitish  = 'test_branch')
+        # local_release is of an inappropriate type 
+        test_args = copy.deepcopy(args)
+        test_args['local_release'] = 1
+        with self.assertRaises(AttributeError):
+            tools.release(**test_args)
 
-        # The local_release argument currently has to specify a /release/ directory.
-        # Capitalisation matters!
-        with self.assertRaises(ReleaseError), nostderrout():
-            tools.release('test_version', 
-                          DriveReleaseFiles = ['paper.pdf', 'plot.pdf'], 
-                          local_release     = 'root/Release/folder', 
-                          org               = 'org', 
-                          repo              = 'repo', 
-                          target_commitish  = 'test_branch')   
+        # local_release doesn't contain a /release/ subdirectory. Caps matter!
+        test_args = copy.deepcopy(args)
+        test_args['local_release'] = 'root/Release/folder'      
+        with self.assertRaises(ReleaseError):
+            tools.release(**test_args)   
 
-        # When org/repo refers to a nonexistent GitHub repository, the  
-        # release request will fail. This should raise a requests error.
-        with self.assertRaises(requests.exceptions.HTTPError), nostderrout():
-            tools.release('test_version', 
-                          DriveReleaseFiles = ['paper.pdf', 'plot.pdf'], 
-                          local_release     = 'root/release', 
-                          org               = 'orgg', # Misspelling
-                          repo              = 'repo', 
-                          target_commitish  = 'test_branch')  
+        # org/repo refers to a nonexistent GitHub repository
+        test_args = copy.deepcopy(args)
+        test_args['org'] = 'orgg' # Misspelling
+        with self.assertRaises(requests.exceptions.HTTPError):
+            tools.release(**test_args)  
+        
+        test_args = copy.deepcopy(args)
+        test_args['repo'] = 1 # Incorrect type
+        with self.assertRaises(requests.exceptions.HTTPError):
+            tools.release(**test_args)  
 
-        with self.assertRaises(requests.exceptions.HTTPError), nostderrout():
-            tools.release('test_version', 
-                          DriveReleaseFiles = ['paper.pdf', 'plot.pdf'], 
-                          local_release     = 'root/release', 
-                          org               = 'org', 
-                          repo              = 1, # Wrong type
-                          target_commitish  = 'test_branch')  
+        # DriveReleaseFiles is non-container, non-string that bool() 
+        # evaluates as True
+        test_args = copy.deepcopy(args)
+        test_args['DriveReleaseFiles'] = True         
+        with self.assertRaises(TypeError):
+            tools.release(**test_args)    
 
-        # Passing a non-container, non-string value for which bool() returns True to DriveReleaseFiles
-        # raises a TypeError
-        with self.assertRaises(TypeError), nostderrout():
-            tools.release('test_version', 
-                          DriveReleaseFiles = True, 
-                          local_release     = 'root/release/repo-test_branch', 
-                          org               = 'org', 
-                          repo              = 'repo', 
-                          target_commitish  = 'test_branch')    
-
-        # Passing a container holding non-string objects to DriveReleaseFiles
-        # raises an AttributeError
-        with self.assertRaises(AttributeError), nostderrout():
-            tools.release('test_version', 
-                          DriveReleaseFiles = [1, 2, 3], 
-                          local_release     = 'root/release/repo-test_branch', 
-                          org               = 'org', 
-                          repo              = 'repo', 
-                          target_commitish  = 'test_branch')  
+        # DriveReleaseFiles holds non-strings
+        test_args = copy.deepcopy(args)
+        test_args['DriveReleaseFiles'] = [1, 2, 3]         
+        with self.assertRaises(AttributeError):
+            tools.release(**test_args)           
 
 
 if __name__ == '__main__':
